Link all pages together for easy navigation
I linked all the pages together with <a href>.
I had some trouble linking to the index.html,
because my folder structure was different from
the one on TOP because i think itll make it
easier to maintain this way. But i figured out
that i have to go two folders up with "../../".